+++ haze-blur/src/androidMain/kotlin/dev/chrisbanes/haze/blur/BlurVisualEffect.android.kt
@@ -11,20 +11,19 @@ import dev.chrisbanes.haze.VisualEffectContext
 internal actual fun BlurVisualEffect.updateDelegate(
   context: VisualEffectContext,
   drawScope: DrawScope,
-) {
-  val canUseRenderEffect = Build.VERSION.SDK_INT >= 31 &&
-    drawScope.drawContext.canvas.nativeCanvas.isHardwareAccelerated
+): BlurVisualEffect.Delegate {
+  val canUseRenderEffect = canUseRenderEffect(
+    sdkInt = Build.VERSION.SDK_INT,
+    isHardwareAccelerated = drawScope.drawContext.canvas.nativeCanvas.isHardwareAccelerated,
+  )
 
   val blurEnabled = blurEnabled
 
   if (blurEnabled && canUseRenderEffect) {
-    val newBlurEffect = when (delegate) {
+    return when (delegate) {
       is RenderEffectBlurVisualEffectDelegate -> delegate
       else -> RenderEffectBlurVisualEffectDelegate(this)
     }
-    // We have a valid blur effect, so return
-    delegate = newBlurEffect
-    return
   }
 
   if (blurEnabled) {
@@ -33,14 +32,14 @@ internal actual fun BlurVisualEffect.updateDelegate(
       else -> RenderScriptBlurVisualEffectDelegate.createOrNull(this, context)
     }
     if (newDelegate != null) {
-      // We have a valid blur effect, so return
-      delegate = newDelegate
-      return
+      return newDelegate
     }
   }
 
   // If we reach here, this is the fallback case of using a scrim
-  if (delegate !is ScrimBlurVisualEffectDelegate) {
-    delegate = ScrimBlurVisualEffectDelegate(this)
+  return if (delegate is ScrimBlurVisualEffectDelegate) {
+    delegate
+  } else {
+    ScrimBlurVisualEffectDelegate(this)
   }
 }